requirements.txt
fastapi
uvicorn
openpyxl
python-multipart
pandas
numpy
scikit-learn
httpx
openai
matplotlib
python-dotenv